isReachable implementation to iOS app
debugging with isReachable method

+++ shared/src/commonMain/kotlin/id/tru/kmm/phonecheckexample/KReachabilityError.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class KReachabilityError (
     val type: String?,
     val title: String?,
-    val status: Int,
+    val status: Long,
     val detail: String?,
     ) {
